resi

+++ ms-productos/src/main/java/com/example/msproductos/repository/ProductoRepository.java
@@ -3,5 +3,6 @@ package com.example.msproductos.repository;
 import com.example.msproductos.entity.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductoRepository extends JpaRepository<Producto, Long> {
+public interface ProductoRepository extends
+        JpaRepository<Producto, Integer> {
 }
